+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 	- Renamed "z88dkMapFile" to "mapFile".
 	- "addOffset" removed for z88dk. Use "mapFile" instead.
 	- "mapFile" is mandatory now.
+- launch.json: "disassemblerArgs"/"numberOfLines": New parameter to control the number of displayed disassembled lines.
 
 # 1.4.9
 - Fixed issue #29: Zsim: load instruction not executed properly for addresses between 0x0000 and 0x3FFF